apple-codesign: set __LINKEDIT segment vmsize to nearest 16kb boundary
Before, when adjusting the __LINKEDIT segment size, we only updated the
`filesize` field. If the new `filesize` were larger than the old
`vmsize` field, this would constitute an invalid segment instruction
since there wouldn't be enough memory reserved for the segment data.

Because Apple's codesign appears to round the `vmsize` up to the next
16kb boundary, this wasn't a problem for many binaries. However, if
the changed size was just right, it could result in a malformed Mach-O
binary.

This commit updates the __LINKEDIT segment's vmsize so that it rounds
up to the next 16kb boundary. This ensures vmsize >= filesize.

Closes #514.